chore: move to shared + remove version from UA
