+++ shared/lib/get-wm-endpoints.js
