+++ shared/lib/microformat/dom.js
