+++ shared/lib/passport.js
