+++ shared/lib/rss/is.js
