+++ shared/lib/send/index.js
